8/11 Bare Bones Dialogue stuff
I've made a working dialogue system that works well enough that I can improve upon it so that it works in the way I intend it to work. So far it only works with a single NPC and doesnt reset, which are things to work on in the future. At the very least it moves on and off the screen when dialogue starts and stops, along with typing out the text I input into the NPC and displaying it's name. All in all, I think this was a generally successful bit of coding. Good job me. ;P

-NF