8/11 Night Cleanup
Cleaned up some of the scripts that I made last night. I got rid of a lot of the code in favor of shorter less Update reliant code. Also mesed with the position of the Dialogue icon. it now shows up in the corner of the screen without clipping outside of it. I still dont really like the look of it. I'll bring it up with Tim later. it works as a placeholder and a concept. Welp that's all I got tonight. Till next time.

-NF