12/3 Transition work and breaking the canvas
I made a transition between scenes but unofrtunately it broke the canvas. I need to change some things with the dialogue system so that the instead of using setactive I use camera culling to make the canvas not appear. Also I may need to make a separate canvas for the HUD and Dialogue.